+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- Add location methods: `driver.device.getLocation`, `driver.device.setLocation`
 
 ## [0.0.7] - 2019-07-07
 - Add get sessions command: `driver.sessions.get()`